+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Pending
 - Admin Manual Payment:
-  - Feat: Show confirmation email template dropdown only when "Send confirmation email" is checked
+  - Show confirmation email template dropdown only when "Send confirmation email" is checked
+  - Create endpoint "responses/{responseId}/email" which sends out a confirmation email / receipt. This can now be accessed by checking the "Send confirmation email" checkbox.
 - Fix: IPN handlers not working with non-ascii characters (#204)
 - Fix bug in which responses table would error when no confirmationEmailTemplates are defined